refactor: hardware_models -> hardware_model
