+++ src/sinsei_umiusi_control/hardware_model/indicator_led_model.cpp
@@ -1,11 +1,11 @@
-#include "sinsei_umiusi_control/hardware_models/indicator_led_model.hpp"
+#include "sinsei_umiusi_control/hardware_model/indicator_led_model.hpp"
 
 #include <memory>
 
 #include "sinsei_umiusi_control/util/gpio.hpp"
 
 namespace suc = sinsei_umiusi_control;
-namespace suchm = suc::hardware_models;
+namespace suchm = suc::hardware_model;
 
 suchm::IndicatorLedModel::IndicatorLedModel(std::shared_ptr<util::GpioWrapper> gpio)
 : gpio(std::move(gpio)) {}
